working and tested changes of solr 9.1

+++ solr/core/src/java/org/apache/solr/core/MMapDirectoryFactory.java
@@ -37,20 +37,20 @@ import org.slf4j.LoggerFactory;
  *   <li>unmap -- See {@link MMapDirectory#setUseUnmap(boolean)}
  *   <li>preload -- See {@link MMapDirectory#setPreload(boolean)}
  *   <li>maxChunkSize -- The Max chunk size. See {@link MMapDirectory#MMapDirectory(Path,
- *       LockFactory, long)}
+ *       LockFactory, int)}
  * </ul>
  */
 public class MMapDirectoryFactory extends StandardDirectoryFactory {
   private static final Logger log = LoggerFactory.getLogger(MethodHandles.lookup().lookupClass());
   boolean unmapHack;
   boolean preload;
-  private long maxChunk;
+  private int maxChunk;
 
   @Override
   public void init(NamedList<?> args) {
     super.init(args);
     SolrParams params = args.toSolrParams();
-    maxChunk = params.getLong("maxChunkSize", MMapDirectory.DEFAULT_MAX_CHUNK_SIZE);
+    maxChunk = params.getInt("maxChunkSize", MMapDirectory.DEFAULT_MAX_CHUNK_SIZE);
     if (maxChunk <= 0) {
       throw new IllegalArgumentException("maxChunk must be greater than 0");
     }
